+++ src/text/index.ts
@@ -1,3 +1,3 @@
 export { chunkByTokenCount, chunkBySentence } from './chunking';
-export { getTokenizerInfo } from './tokenizer';
+export { getTokenizerInfo, createTokenizer } from './tokenizer';
 export { chunkByStructure } from './markdown';